requirements.txt
smolagents[gradio]
requests
python-dotenv